chore: updating unittests
- Moved test files into the correct directory

TO-DO: Fix input validation test
    - Make a mock object for the GUI
